WIP: Update HiveRunner to use Hive 3.0.0
This commit produces a version of HiveRunner that works with Hive 3.0.0
All tests run and pass, with the sole exception of
PreV200HiveShellHiveCliEmulationTest.

This commit is a starting point for 3.0.0 support, but it takes a
pretty blunt approach in updating HiveRunner.  I've only made those
changes strictly necessary to compile and pass tests on the updated Hive
version, and I did so without tons of context on Hive 3 itself.  There
might be changes needed, or new Hive3 features that should be tested,
or...

Changes made include:
- update pom dependencies for newer Hive/Hadoop versions
- alter packaging to create a shaded uber-jar (Hadoop2 and its
  transitive deps are still widely used, and lag behind Hadoop3's
  transitive deps quite a bit.  Shading our use of Hadoop3 and some deps
  eases integration with other projects.
- alter 'HiveConf' initialization to account for Metastore configuration
  changes in Hive 3